Move everything to convert-hf-to-gguf.py
